refactor: enlarge step clickable area
