+++ packages/steps/src/style/content.tsx
@@ -1,6 +1,5 @@
 // content that contains title and description
 import { css } from "@emotion/react"
-import { globalColor, illaPrefix } from "@illa-design/theme"
 import { SerializedStyles } from "@emotion/serialize"
 import { StepVariant, LabelPlacement } from "../interface"
 import { isVerticalLabel } from "../style"
@@ -9,17 +8,14 @@ export function applyContentStyle({
   variant,
   labelPlacement,
   direction,
-  hoverable,
 }: {
   variant: StepVariant
   labelPlacement: LabelPlacement
   direction: LabelPlacement
-  hoverable: boolean
 }): SerializedStyles {
   return css([
     applyContentDisplay({ variant, labelPlacement, direction }),
     applyContentSize({ variant, labelPlacement, direction }),
-    applyHover(hoverable),
   ])
 }
 
@@ -72,32 +68,4 @@ function applyContentSize({
   }
 
   return css([width, margin])
-}
-
-function applyHover(hoverable: boolean): SerializedStyles {
-  if (!hoverable) {
-    return css``
-  }
-
-  const hoverColor = globalColor(`--${illaPrefix}-blue-05`)
-
-  return css`
-    & > * {
-      transition: all 0.15s ease-in-out;
-      &:first-child {
-        transition-duration: 0.2s;
-      }
-    }
-
-    &:hover > * {
-      transition: all 0.2s ease-in-out;
-
-      &:first-child {
-        transition-duration: 0.15s;
-      }
-
-      cursor: pointer;
-      color: ${hoverColor};
-    }
-  `
 }
