+++ packages/steps/src/style/wrapper.tsx
@@ -1,6 +1,7 @@
 // wrapper container that contains icon & title & description
 import { css } from "@emotion/react"
 import { SerializedStyles } from "@emotion/serialize"
+import { globalColor, illaPrefix } from "@illa-design/theme"
 import { StepVariant, LabelPlacement, StepStatus } from "../interface"
 import { statusColor, isVerticalLabel } from "../style"
 
@@ -10,12 +11,14 @@ export function applyWrapperStyle({
   status,
   disabled,
   labelPlacement,
+  hoverable,
 }: {
   direction: LabelPlacement
   variant: StepVariant
   status: StepStatus
   disabled: boolean
   labelPlacement: LabelPlacement
+  hoverable: boolean
 }): SerializedStyles {
   const overflow = isVerticalLabel({ variant, direction, labelPlacement })
     ? "visible"
@@ -60,6 +63,7 @@ export function applyWrapperStyle({
     boxStyle,
     navigactionProcessStatusIndicator,
     applyWrapperCursor(disabled),
+    applyHover(hoverable),
   ])
 }
 
@@ -69,4 +73,36 @@ function applyWrapperCursor(disabled: boolean): SerializedStyles {
         cursor: not-allowed;
       `
     : css``
+}
+
+function applyHover(hoverable: boolean): SerializedStyles {
+  if (!hoverable) {
+    return css``
+  }
+
+  const hoverColor = globalColor(`--${illaPrefix}-blue-03`)
+
+  return css`
+    & > div:last-child > div {
+      transition: all 0.12s ease-in-out;
+      &:first-child {
+        transition-duration: 0.2s;
+      }
+    }
+
+    &:hover {
+      cursor: pointer;
+    }
+
+    &:hover > div:last-child > div {
+      transition: all 0.2s ease-in-out;
+
+      &:first-child {
+        transition-duration: 0.12s;
+      }
+
+      cursor: pointer;
+      color: ${hoverColor};
+    }
+  `
 }
